pyproject.toml
[project]
name = "GreenOps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[litellm]>=0.140.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "fpdf2>=2.8.7",
    "pandas>=2.3.3",
    "uuid>=1.30",
]
